+++ build.sh
@@ -1,0 +1,2 @@
+javac Tile.java
+javac Minesweeper.java