+++ Changelog.md
@@ -1,1 +1,25 @@
-#Changelog
+# Changelog
+
+
+## Version 0.1.1
+### Added
+
+#### Minesweeper.java
+- The constructor initializes the row, column, isCovered, and tileDisplayValue of all tiles correctly
+- The display method displays the Minesweeper grid
+- The while loop in main(String[] args) displays the user interface that will be used
+
+
+#### Tile.java
+- All tiles have a row, column, hasBomb, isCovered, tileDisplayValue, and adjacentBombs values
+- Setters and Getters for the tileDisplayValue
+
+- .gitignore added
+
+
+
+
+## Version 0.0.1
+### Added 
+- Changelog.md first created
+- readme.md first created